Moved contrib.gis.geometry.regex to contrib.gis.geometry.
