+++ django/contrib/gis/geometry.py
